backend/requirements.txt
fastapi
uvicorn
requests
pydantic
pydantic-settings
python-dotenv
langchain
langchain_google_genai
google-generativeai
langchain_openai
langchain_anthropic
langchain-text-splitters
pgvector
slowapi
pydantic[email]
sqlalchemy==2.0.23
psycopg2-binary
alembic==1.13.0
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0
python-multipart
pandas
werkzeug